Renamed namespace General to Constants
