+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Blocks/PersistedQueryGraphiQLBlock.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Blocks;
 
-use GraphQLAPI\GraphQLAPI\General\RequestParams;
+use GraphQLAPI\GraphQLAPI\Constants\RequestParams;
 use GraphQLAPI\GraphQLAPI\Services\Helpers\EditorHelpers;
 use PoP\ComponentModel\Facades\Instances\InstanceManagerFacade;
