+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Clients/CustomEndpointClientTrait.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Clients;
 
-use GraphQLAPI\GraphQLAPI\General\RequestParams;
+use GraphQLAPI\GraphQLAPI\Constants\RequestParams;
 use GraphQLAPI\GraphQLAPI\PostTypes\GraphQLEndpointPostType;
 use PoP\ComponentModel\Misc\RequestUtils;
