+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Constants/BlockConstants.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace GraphQLAPI\GraphQLAPI\General;
+namespace GraphQLAPI\GraphQLAPI\Constants;
 
 class BlockConstants
 {
