+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/ContentProcessors/AbstractContentParser.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\ContentProcessors;
 
-use GraphQLAPI\GraphQLAPI\General\RequestParams;
+use GraphQLAPI\GraphQLAPI\Constants\RequestParams;
 use GraphQLAPI\GraphQLAPI\PluginConstants;
 use GraphQLAPI\GraphQLAPI\Services\Helpers\LocaleUtils;
 use InvalidArgumentException;
