+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/PostTypes/AbstractGraphQLQueryExecutionPostType.php
@@ -7,7 +7,7 @@ namespace GraphQLAPI\GraphQLAPI\PostTypes;
 use WP_Post;
 use PoP\ComponentModel\State\ApplicationState;
 use GraphQLAPI\GraphQLAPI\Services\Helpers\BlockHelpers;
-use GraphQLAPI\GraphQLAPI\General\RequestParams;
+use GraphQLAPI\GraphQLAPI\Constants\RequestParams;
 use GraphQLAPI\GraphQLAPI\PostTypes\AbstractPostType;
 use GraphQLAPI\GraphQLAPI\Blocks\SchemaConfigurationBlock;
 use PoP\ComponentModel\Facades\Instances\InstanceManagerFacade;
