+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/Helpers/MenuPageHelper.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\Services\Helpers;
 
-use GraphQLAPI\GraphQLAPI\General\RequestParams;
+use GraphQLAPI\GraphQLAPI\Constants\RequestParams;
 
 /**
  * About menu page
