+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/Services/MenuPages/AbstractDocAboutMenuPage.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace GraphQLAPI\GraphQLAPI\Services\MenuPages;
 
 use InvalidArgumentException;
-use GraphQLAPI\GraphQLAPI\General\RequestParams;
+use GraphQLAPI\GraphQLAPI\Constants\RequestParams;
 use GraphQLAPI\GraphQLAPI\Facades\ContentProcessors\MarkdownContentParserFacade;
 
 /**
